Handle empty input structures in Validator#validateSyntax
- Allows for empty segments to be validated properly
- Re-factor tests, add test for empty segment that contains element
