+++ src/main/java/io/xlate/edi/internal/stream/validation/Validator.java
@@ -898,7 +898,8 @@ public class Validator {
         final int componentIndex = location.getComponentPosition() - 1;
         final List<UsageNode> children = structure.getChildren();
 
-        for (int i = index, max = children.size(); i < max; i++) {
+        // Ensure the start index is at least zero. Index may be -1 for empty segments
+        for (int i = Math.max(index, 0), max = children.size(); i < max; i++) {
             if (isComposite) {
                 location.incrementComponentPosition();
             } else {
